content/media/test/test_bug919265.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=919265
-->
<head>
  <meta charset='utf-8'>
  <title>Regression test for bug 919265 - Leak on VTTCue::GetCueAsHTML()</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

// We shouldn't leak upon shutdown.
(new VTTCue(0, 0, "")).getCueAsHTML();

// We need to assert something for Mochitest to be happy.
ok(true);
SimpleTest.finish();
</script>
</pre>
</body>
</html>
